# swagger_server/models/sbf_res.py
# coding: utf-8
# type: ignore
# pylint: disable-all
from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class SBFRes(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, text: str = None, close_context: str = None):  # noqa: E501
        """SBFRes - a model defined in Swagger

        :param text: The text of this SBFRes.  # noqa: E501
        :type text: str
        :param close_context: The close_context of this SBFRes.  # noqa: E501
        :type close_context: str
        """
        self.swagger_types = {
            'text': str,
            'close_context': str
        }

        self.attribute_map = {
            'text': 'text',
            'close_context': 'closeContext'
        }

        self._text = text
        self._close_context = close_context

    @classmethod
    def from_dict(cls, dikt) -> 'SBFRes':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The SBFRes of this SBFRes.  # noqa: E501
        :rtype: SBFRes
        """
        return util.deserialize_model(dikt, cls)

    @property
    def text(self) -> str:
        """Gets the text of this SBFRes.


        :return: The text of this SBFRes.
        :rtype: str
        """
        return self._text

    @text.setter
    def text(self, text: str):
        """Sets the text of this SBFRes.


        :param text: The text of this SBFRes.
        :type text: str
        """

        self._text = text

    @property
    def close_context(self) -> str:
        """Gets the close_context of this SBFRes.


        :return: The close_context of this SBFRes.
        :rtype: str
        """
        return self._close_context

    @close_context.setter
    def close_context(self, close_context: str):
        """Sets the close_context of this SBFRes.


        :param close_context: The close_context of this SBFRes.
        :type close_context: str
        """

        self._close_context = close_context